refactor: move swagger tasks to a different file
